backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "brave-search>=0.1.8",
    "fastapi>=0.104.0",
    "google-genai>=1.2.0",
    "ipykernel>=6.29.5",
    "marimo>=0.13.15",
    "markdownify>=1.1.0",
    "mirascope[anthropic,google,openai]>=1.25.0",
    "pip>=25.1.1",
    "pydantic>=2.11.5",
    "python-dotenv>=1.1.0",
    "ruff>=0.11.13",
    "uvicorn>=0.24.0",
]
